+++ CHANGELOG.md
@@ -14,6 +14,7 @@ x.y.z / 2014-06-dd
 * changed method signature of `decode(version, base58str)` to `decode(base58str, version)`
 * added method `createValidator()`
 * changed method signature of `validate(version, base58str)` to `validate(base58str, version)`
+* renamed method `validate()` to `isValid()`
 
 0.2.0 / 2014-03-10
 ------------------